Solved Day 1 part 2!
Wrestled with func syntax for a bit. Was trying to declare modes. Derp.

```
> time ./run-day 1
../src/day1.m:068: In clause for predicate `main'/2:
../src/day1.m:068:   warning: variables `X, Y' each have overlapping scopes.
Since 456 + 1564 = 2020, the answer to part 1 is that 456 * 1564 = 713184.
Part 2:
["399*764*857 = 261244452", "399*857*764 = 261244452", "764*399*857 = 261244452", "764*857*399 = 261244452", "857*399*764 = 261244452", "857*764*399 = 261244452"]

________________________________________________________
Executed in  712.14 millis    fish           external
   usr time  544.40 millis  120.00 micros  544.28 millis
   sys time  154.78 millis  671.00 micros  154.11 millis
```

Not sure what the "overlapping scope" bit is about, or why the Triples don't
have that problem.

I was unwilling to require all numbers to be unequal or better yet sorted,
in case there were dupes (we could have had 1010 and 1010!),
but that led to there being multiple of the same answer in different
orders.

Probably there's a way to short-circuit the solving using either
multi_cc (committed choice) or something else in the solutions module.