Solved day 19 part 2
Gotcha: Don't use relative backref - there are groups inside the
expression we're expanding! Instead use a named backref.

pcre2grep -e '^((((a(b(a(aa|ba)|bab)|a(bbb|(b(b|a)|aa)a))|b((a(aa|bb)|b(b|a)(b|a))b|((aa|bb)b|((b|a)a|ab)a)a))a|(((a(aa|ab)|bba)a|((bb|ba)b|(b(b|a)|aa)a)b)a|(a((ab|b(b|a))a|(b|a)(b|a)b)|b(aab|b(ab|bb)))b)b)a|(a(((a(aa|(b|a)b)|b(aa|ba))b|(b(aa|ba)|a(ba|ab))a)a|(a(a(aa|ba)|b(aa|ab))|b((ba|ab)a|(aa|(b|a)b)b))b)|b(a(((ab|bb)b|(ba|ab)a)a|((ab|bb)b|(aa|ba)a)b)|b(b(aba|(ab|bb)b)|a(bab|(aa|bb)a))))b)b|(b(b((((ab|b(b|a))a|(ba|ab)b)b|(a(ab|bb)|b(ab|b(b|a)))a)a|(a(a(b(b|a)|aa)|bbb)|b(a(aa|ba)|b(aa|(b|a)b)))b)|a((a(a(b|a)(b|a)|b(ab|b(b|a)))|b((ab|bb)a|abb))b|((abb|b(aa|ba))b|(bba|aab)a)a))|a(((bb(ba|ab)|a(a(aa|ba)|bab))a|(a(a(aa|ba)|bba)|b(a(b|a)(b|a)|b(ba|ab)))b)a|(a(a((b(b|a)|aa)b|baa)|b((ab|bb)b|(aa|ba)a))|b(((aa|ba)a|bbb)b|bbaa))b))a)+(?P<Eleven>((((a(b(a(aa|ba)|bab)|a(bbb|(b(b|a)|aa)a))|b((a(aa|bb)|b(b|a)(b|a))b|((aa|bb)b|((b|a)a|ab)a)a))a|(((a(aa|ab)|bba)a|((bb|ba)b|(b(b|a)|aa)a)b)a|(a((ab|b(b|a))a|(b|a)(b|a)b)|b(aab|b(ab|bb)))b)b)a|(a(((a(aa|(b|a)b)|b(aa|ba))b|(b(aa|ba)|a(ba|ab))a)a|(a(a(aa|ba)|b(aa|ab))|b((ba|ab)a|(aa|(b|a)b)b))b)|b(a(((ab|bb)b|(ba|ab)a)a|((ab|bb)b|(aa|ba)a)b)|b(b(aba|(ab|bb)b)|a(bab|(aa|bb)a))))b)b|(b(b((((ab|b(b|a))a|(ba|ab)b)b|(a(ab|bb)|b(ab|b(b|a)))a)a|(a(a(b(b|a)|aa)|bbb)|b(a(aa|ba)|b(aa|(b|a)b)))b)|a((a(a(b|a)(b|a)|b(ab|b(b|a)))|b((ab|bb)a|abb))b|((abb|b(aa|ba))b|(bba|aab)a)a))|a(((bb(ba|ab)|a(a(aa|ba)|bab))a|(a(a(aa|ba)|bba)|b(a(b|a)(b|a)|b(ba|ab)))b)a|(a(a((b(b|a)|aa)b|baa)|b((ab|bb)b|(aa|ba)a))|b(((aa|ba)a|bbb)b|bbaa))b))a)(?P>Eleven)?(((a(b(((bb|ba)a|(aa|(b|a)b)b)a|(b|a)(aa|ba)b)|a((((b|a)a|ab)b|aba)a|(bba|aab)b))|b(b(a(baa|abb)|b(aaa|(aa|ab)b))|a(b((ba|ab)a|(aa|(b|a)b)b)|a((aa|ab)b|bba))))a|((((aba|(ab|bb)b)b|(b((b|a)a|ab)|aab)a)b|(((aa|ba)a|bbb)a|((ab|bb)a|(b|a)(b|a)b)b)a)a|(((b((b|a)a|ab)|a(aa|bb))a|(a(aa|ab)|bba)b)b|(b((bb|ba)a|(aa|(b|a)b)b)|a(b(aa|bb)|abb))a)b)b)b|(a(b((b(aa|bb)(b|a)|a(b|a)(aa|ba))a|(a((ab|b(b|a))b|(b|a)(b|a)a)|b((ba|ab)b|(aa|(b|a)b)a))b)|a(a(b(bab|(bb|ba)a)|a(b(b|a)(b|a)|aab))|b(((aa|(b|a)b)b|(b(b|a)|aa)a)a|(b|a)(ba|ab)b)))|b((b((a(aa|ba)|b(aa|ab))b|(aba|(ab|bb)b)a)|a(b(b(aa|(b|a)b)|a(aa|ab))|abaa))b|(b(((aa|bb)b|aaa)b|(a(aa|ba)|baa)a)|a(((aa|ab)a|(b(b|a)|aa)b)b|(a(b(b|a)|aa)|b(aa|(b|a)b))a))a))a))$' input/day19.txt | wc -l
     346